iam/beginner/cloud-storage/src/backend/requirements.txt
Django==6.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
boto3==1.34.34
python-dotenv==1.0.0